local/requirements.txt
flask
numpy
pandas
scikit-learn
sklearn-pandas
joblib
shap
gunicorn
